Commentaires et ménage

+++ src/init.h
@@ -13,7 +13,7 @@
 
 int16_t checkInstallationPaths(char *base_path, int16_t try_to_create_flag);
 int16_t checkParamsDb(char *sqlite3_db_param_path, int16_t *cause);
-int create_configs_php(char *gui_home, char *params_db_fullname, char *php_log_fullname);
+int16_t create_configs_php(char *gui_home, char *params_db_fullname, char *php_log_fullname);
 int16_t initMeaEdomus(int16_t mode, char **params_list, char **keys);
 int16_t updateMeaEdomus(char **params_list, char **keys);
